+++ R/pool.r
@@ -26,9 +26,11 @@ Pool = R6::R6Class("Pool",
 
         info = function() {
             info = private$master$list_workers()
-            times = do.call(rbind, info$time)[,1:3]
+            times = do.call(rbind, info$time)[,1:3,drop=FALSE]
             mem = function(field) sapply(info$mem, function(m) sum(m[,field] * c(56,1)))
-            do.call(data.frame, c(info[c("worker", "status")], as.data.frame(times),
+            do.call(data.frame, c(info[c("worker", "status")],
+                                  current=list(info$worker==info$cur),
+                                  info["calls"], as.data.frame(times),
                                   list(mem.used=mem("used"), mem.max=mem("max used"))))
         },
 
@@ -93,8 +95,8 @@ Pool = R6::R6Class("Pool",
             private$master$send(Sys.sleep(wait/1000))
         },
 
-        recv = function() {
-            private$master$recv(-1L)
+        recv = function(timeout=-1L) {
+            private$master$recv(timeout)
         },
 
         cleanup = function(timeout=5000) {